refactored components into different files
